% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{knn}
\alias{knn}
\title{Get Nearest Neighbors}
\usage{
knn(rTree, y, k)
}
\arguments{
\item{rTree}{An \link{RTree} object.}

\item{y}{A 2-column numeric matrix of point coordinates.}

\item{k}{A positive integer.}
}
\description{
For each point \eqn{y_i} in set \code{y}, returns the row-indices of the \code{k} points indexed in \code{rTree}
that are closest to \eqn{y_i}.
}
