% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RTree}
\alias{RTree}
\title{Create an RTree}
\usage{
RTree(x)
}
\arguments{
\item{x}{A 2-column numeric matrix of point coordinates.}
}
\value{
An RTree S3 object.
}
\description{
Organizes points in an R-tree.
}
\details{
The R-tree is created using the quadratic splitting algorithm, with the maximum number of elements
per node set to 16. See \url{http://www.boost.org/doc/libs/1_63_0/libs/geometry/doc/html/geometry/spatial_indexes/introduction.html} for details.
}
