% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos.R
\name{rtree_geos}
\alias{rtree_geos}
\alias{rtree_geos_insert}
\alias{rtree_geos_query}
\alias{rtree_geos_query_once}
\alias{rtree_geos_size}
\title{GEOS Tree}
\usage{
rtree_geos(x, node_capacity = 16)

rtree_geos_insert(tree, x)

rtree_geos_query(tree, x)

rtree_geos_query_once(tree, x)

rtree_geos_size(tree)
}
\arguments{
\item{x}{A vector of geometries for which \code{\link[wk:wk_bbox]{wk::wk_envelope()}} works}

\item{node_capacity}{Node size}

\item{tree}{A tree index object}
}
\description{
GEOS Tree
}
\examples{
points <- wk::xy(runif(10), runif(10))
tree <- rtree_geos(points)
rtree_geos_query(tree, wk::rct(0.25, 0.25, 0.75, 0.75))
rtree_geos_size(tree)

}
