% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{countWithinDistance}
\alias{countWithinDistance}
\title{Count Points Within Distance}
\usage{
countWithinDistance(rTree, y, distance)
}
\arguments{
\item{rTree}{An \link{RTree} object.}

\item{y}{A 2-column numeric matrix of point coordinates.}

\item{distance}{A positive scalar.}
}
\value{
An integer vector of with one entry for each row of `y`.
}
\description{
For each point \eqn{y_i} in set \code{y}, returns the number of
points indexed in \code{rTree}
that are within a given \code{distance} of \eqn{y_i}.
}
\examples{
# For two sets of points, A and B, count points in B within
# a distance of 0.05 to points in A.
n <- 10^4
A <- cbind(runif(n), runif(n))
B <- cbind(runif(n), runif(n))

# Make an RTree on B
B_rtree <- RTree(B)

# Find points of B within 0.05 of each point in A
counts <- countWithinDistance(B_rtree, A, 0.05)

# Each element of counts is a count of points in B
print(counts[[1]])
}
