% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtree-package.R
\docType{package}
\name{rtree-package}
\alias{rtree}
\alias{rtree-package}
\title{rtree: Experimental 'RTree' Spatial Indexing}
\description{
Experimental implementations of various 'RTree' spatial indexing.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
