% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatgeobuf-packed.R
\name{rtree_flatgeobuf}
\alias{rtree_flatgeobuf}
\alias{rtree_flatgeobuf_query}
\title{FlatGeobuf PackedRTree}
\usage{
rtree_flatgeobuf(x, node_capacity = 16)

rtree_flatgeobuf_query(tree, x)
}
\arguments{
\item{x}{A vector of geometries for which \code{\link[wk:wk_bbox]{wk::wk_envelope()}} works}

\item{node_capacity}{Node size}

\item{tree}{A tree index object}
}
\description{
FlatGeobuf PackedRTree
}
\examples{
points <- wk::xy(runif(10), runif(10))
tree <- rtree_flatgeobuf(points)
rtree_flatgeobuf_query(tree, wk::rct(0.25, 0.25, 0.75, 0.75))

}
