% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{withinDistance}
\alias{withinDistance}
\title{Get Points Within Distance}
\usage{
withinDistance(rTree, y, distance)
}
\arguments{
\item{rTree}{An \link{RTree} object.}

\item{y}{A 2-column numeric matrix of point coordinates.}

\item{distance}{A positive scalar.}
}
\description{
For each point \eqn{y_i} in set \code{y}, returns the row-indices of the points indexed in \code{rTree}
that are within a given \code{distance} of \eqn{y_i}.
}
